// web/src/api/generated/models/NotificationChannel.ts
/* tslint:disable */
/* eslint-disable */
/**
 * eimir.
 * Application Core. OpenAPI is the authoritative contract.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const NotificationChannel = {
    IN_APP: 'IN_APP',
    PUSH: 'PUSH',
    EMAIL: 'EMAIL'
} as const;
export type NotificationChannel = typeof NotificationChannel[keyof typeof NotificationChannel];


export function instanceOfNotificationChannel(value: any): boolean {
    for (const key in NotificationChannel) {
        if (Object.prototype.hasOwnProperty.call(NotificationChannel, key)) {
            if (NotificationChannel[key as keyof typeof NotificationChannel] === value) {
                return true;
            }
        }
    }
    return false;
}

export function NotificationChannelFromJSON(json: any): NotificationChannel {
    return NotificationChannelFromJSONTyped(json, false);
}

export function NotificationChannelFromJSONTyped(json: any, ignoreDiscriminator: boolean): NotificationChannel {
    return json as NotificationChannel;
}

export function NotificationChannelToJSON(value?: NotificationChannel | null): any {
    return value as any;
}

export function NotificationChannelToJSONTyped(value: any, ignoreDiscriminator: boolean): NotificationChannel {
    return value as NotificationChannel;
}